basic search bar plus DB fix(migration file when i was adding a column broke the DB)

+++ db/migrate/20200324221936_create_carsforsales.rb
@@ -4,7 +4,7 @@ class CreateCarsforsales < ActiveRecord::Migration[5.2]
       t.string :title
       t.string :brand
       t.integer :year
-      t.integer :engine
+      t.decimal :engine
       t.string :transmission
       t.integer :mileage
       t.string :color
